Import subprocess.call if run is not available
The subprocess.run function was added in Python 3.5. Prior to this, the
subprocess.call function was the high level subprocess API. In order to
support Python versions before 3.5, a backup import is added that
imports the call function as run. The rest of the code can then remain
the same.